Fix BinaryClassifier

+++ src/main/java/visrec/djl/ml/dataset/CsvDataset.java
@@ -29,7 +29,7 @@ public class CsvDataset extends RandomAccessDataset {
 
     @Override
     public Record get(NDManager manager, long index) {
-        CSVRecord record = records.get((int) index);
+        CSVRecord record = records.get(Math.toIntExact(index));
         int size = record.size();
         float[] data = new float[size - 1];
         for (int i = 0; i < size - 1; ++i) {
